* Removed old "MainCart.js" and changed "HanokCart" to "MainCart". * App creates cart on customer login instead of when clicking restaurant. * Changed maximum distance for "Filter" and changed step to 5km.
